long double

+++ src/types/Mat3d.cpp
@@ -16,9 +16,9 @@ Mat3d::Mat3d() {
             m[i][j] = 0.0;
 }
 
-Mat3d::Mat3d(double m00, double m01, double m02,
-             double m10, double m11, double m12,
-             double m20, double m21, double m22) {
+Mat3d::Mat3d(long double m00, long double m01, long double m02,
+             long double m10, long double m11, long double m12,
+             long double m20, long double m21, long double m22) {
     m[0][0] = m00; m[0][1] = m01; m[0][2] = m02;
     m[1][0] = m10; m[1][1] = m11; m[1][2] = m12;
     m[2][0] = m20; m[2][1] = m21; m[2][2] = m22;
@@ -41,26 +41,26 @@ Mat3d Mat3d::Identity() {
                  0, 0, 1);
 }
 
-Mat3d Mat3d::Diagonal(double d0, double d1, double d2) {
+Mat3d Mat3d::Diagonal(long double d0, long double d1, long double d2) {
     return Mat3d(d0, 0,  0,
                  0,  d1, 0,
                  0,  0,  d2);
 }
 
 // Операторы доступа
-double& Mat3d::operator()(int i, int j) {
+long double& Mat3d::operator()(int i, int j) {
     return m[i][j];
 }
 
-const double& Mat3d::operator()(int i, int j) const {
+const long double& Mat3d::operator()(int i, int j) const {
     return m[i][j];
 }
 
-double* Mat3d::operator[](int i) {
+long double* Mat3d::operator[](int i) {
     return m[i];
 }
 
-const double* Mat3d::operator[](int i) const {
+const long double* Mat3d::operator[](int i) const {
     return m[i];
 }
 
@@ -94,7 +94,7 @@ Mat3d Mat3d::operator*(const Mat3d& other) const {
     return result;
 }
 
-Mat3d Mat3d::operator*(double scalar) const {
+Mat3d Mat3d::operator*(long double scalar) const {
     Mat3d result;
     for (int i = 0; i < 3; ++i)
         for (int j = 0; j < 3; ++j)
@@ -102,7 +102,7 @@ Mat3d Mat3d::operator*(double scalar) const {
     return result;
 }
 
-Mat3d Mat3d::operator/(double scalar) const {
+Mat3d Mat3d::operator/(long double scalar) const {
     return *this * (1.0 / scalar);
 }
 
@@ -135,14 +135,14 @@ Mat3d& Mat3d::operator*=(const Mat3d& other) {
     return *this;
 }
 
-Mat3d& Mat3d::operator*=(double scalar) {
+Mat3d& Mat3d::operator*=(long double scalar) {
     for (int i = 0; i < 3; ++i)
         for (int j = 0; j < 3; ++j)
             m[i][j] *= scalar;
     return *this;
 }
 
-Mat3d& Mat3d::operator/=(double scalar) {
+Mat3d& Mat3d::operator/=(long double scalar) {
     return *this *= (1.0 / scalar);
 }
 
@@ -164,22 +164,20 @@ Mat3d Mat3d::transpose() const {
     );
 }
 
-// Определитель
-double Mat3d::determinant() const {
+long double Mat3d::determinant() const {
     return m[0][0] * (m[1][1] * m[2][2] - m[1][2] * m[2][1])
          - m[0][1] * (m[1][0] * m[2][2] - m[1][2] * m[2][0])
          + m[0][2] * (m[1][0] * m[2][1] - m[1][1] * m[2][0]);
 }
 
-// Обратная матрица
 Mat3d Mat3d::inverse() const {
-    double det = determinant();
+    long double det = determinant();
     if (std::abs(det) < 1e-15) {
         // Вырожденная матрица
         return Zero();
     }
 
-    double inv_det = 1.0 / det;
+    long double inv_det = 1.0 / det;
 
     return Mat3d(
         (m[1][1] * m[2][2] - m[1][2] * m[2][1]) * inv_det,
@@ -196,45 +194,39 @@ Mat3d Mat3d::inverse() const {
     );
 }
 
-// След матрицы
-double Mat3d::trace() const {
+long double Mat3d::trace() const {
     return m[0][0] + m[1][1] + m[2][2];
 }
 
-// Норма Фробениуса
-double Mat3d::norm() const {
-    double sum = 0.0;
+
+long double Mat3d::norm() const {
+    long double sum = 0.0;
     for (int i = 0; i < 3; ++i)
         for (int j = 0; j < 3; ++j)
             sum += m[i][j] * m[i][j];
     return std::sqrt(sum);
 }
 
-// Спектральная норма (упрощенная версия)
-double Mat3d::spectralNorm() const {
-    // Для точного вычисления нужны собственные значения
-    // Используем оценку через норму Фробениуса
+long double Mat3d::spectralNorm() const {
+
     return norm(); // Это верхняя оценка
 }
 
-// Собственные значения и векторы (упрощенный метод для симметричных матриц)
 bool Mat3d::eigen(Vec3d& eigenvalues, Mat3d& eigenvectors) const {
-    // Упрощенная реализация для симметричных матриц
+
     if (!isSymmetric()) {
-        return false; // Только для симметричных матриц
+        return false;
     }
 
-    // Простой итерационный метод для небольших матриц
     Mat3d A = *this;
     eigenvectors = Mat3d::Identity();
 
     const int max_iterations = 50;
-    const double tolerance = 1e-12;
+    const long double tolerance = 1e-12;
 
     for (int iter = 0; iter < max_iterations; ++iter) {
-        // Находим максимальный внедиагональный элемент
         int p = 0, q = 1;
-        double max_off_diag = std::abs(A(0,1));
+        long double max_off_diag = std::abs(A(0,1));
         for (int i = 0; i < 3; ++i) {
             for (int j = i+1; j < 3; ++j) {
                 if (std::abs(A(i,j)) > max_off_diag) {
@@ -249,17 +241,15 @@ bool Mat3d::eigen(Vec3d& eigenvalues, Mat3d& eigenvectors) const {
             break;
         }
 
-        // Вычисляем вращение Якоби
-        double app = A(p,p);
-        double aqq = A(q,q);
-        double apq = A(p,q);
+        long double app = A(p,p);
+        long double aqq = A(q,q);
+        long double apq = A(p,q);
 
-        double tau = (aqq - app) / (2.0 * apq);
-        double t = (tau >= 0 ? 1.0 : -1.0) / (std::abs(tau) + std::sqrt(1.0 + tau*tau));
-        double c = 1.0 / std::sqrt(1.0 + t*t);
-        double s = t * c;
+        long double tau = (aqq - app) / (2.0 * apq);
+        long double t = (tau >= 0 ? 1.0 : -1.0) / (std::abs(tau) + std::sqrt(1.0 + tau*tau));
+        long double c = 1.0 / std::sqrt(1.0 + t*t);
+        long double s = t * c;
 
-        // Применяем вращение к A
         Mat3d J = Mat3d::Identity();
         J(p,p) = c; J(p,q) = s;
         J(q,p) = -s; J(q,q) = c;
@@ -272,18 +262,15 @@ bool Mat3d::eigen(Vec3d& eigenvalues, Mat3d& eigenvectors) const {
     return true;
 }
 
-// Симметричная часть
 Mat3d Mat3d::symmetricPart() const {
     return (*this + this->transpose()) * 0.5;
 }
 
-// Антисимметричная часть
 Mat3d Mat3d::antisymmetricPart() const {
     return (*this - this->transpose()) * 0.5;
 }
 
-// Проверки
-bool Mat3d::isSymmetric(double tolerance) const {
+bool Mat3d::isSymmetric(long double tolerance) const {
     for (int i = 0; i < 3; ++i) {
         for (int j = i+1; j < 3; ++j) {
             if (std::abs(m[i][j] - m[j][i]) > tolerance) {
@@ -294,7 +281,7 @@ bool Mat3d::isSymmetric(double tolerance) const {
     return true;
 }
 
-bool Mat3d::isZero(double tolerance) const {
+bool Mat3d::isZero(long double tolerance) const {
     for (int i = 0; i < 3; ++i)
         for (int j = 0; j < 3; ++j)
             if (std::abs(m[i][j]) > tolerance)
@@ -302,10 +289,10 @@ bool Mat3d::isZero(double tolerance) const {
     return true;
 }
 
-bool Mat3d::isIdentity(double tolerance) const {
+bool Mat3d::isIdentity(long double tolerance) const {
     for (int i = 0; i < 3; ++i) {
         for (int j = 0; j < 3; ++j) {
-            double expected = (i == j) ? 1.0 : 0.0;
+            long double expected = (i == j) ? 1.0 : 0.0;
             if (std::abs(m[i][j] - expected) > tolerance) {
                 return false;
             }
@@ -314,7 +301,6 @@ bool Mat3d::isIdentity(double tolerance) const {
     return true;
 }
 
-// Вывод
 std::ostream& operator<<(std::ostream& os, const Mat3d& mat) {
     os << "[[" << mat.m[0][0] << ", " << mat.m[0][1] << ", " << mat.m[0][2] << "],\n"
        << " [" << mat.m[1][0] << ", " << mat.m[1][1] << ", " << mat.m[1][2] << "],\n"
@@ -322,13 +308,11 @@ std::ostream& operator<<(std::ostream& os, const Mat3d& mat) {
     return os;
 }
 
-// Внешние операторы
-Mat3d operator*(double scalar, const Mat3d& mat) {
+Mat3d operator*(long double scalar, const Mat3d& mat) {
     return mat * scalar;
 }
 
 Vec3d operator*(const Vec3d& vec, const Mat3d& mat) {
-    // Умножение вектора-строки на матрицу
     return Vec3d(
         vec.x * mat(0,0) + vec.y * mat(1,0) + vec.z * mat(2,0),
         vec.x * mat(0,1) + vec.y * mat(1,1) + vec.z * mat(2,1),
